backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
httpie==2.4.0
idna==2.10
Pillow==8.1.2
psycopg2==2.8.6
Pygments==2.8.1
PySocks==1.7.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
urllib3==1.26.3
